Fixed issue with missing namespaces.
    * Added fix for missing namespaces in classes.
